Move subcommand-related JS to std/cmd/
 - move std/{generate,transform}.ts to std/cmd/, so they aren't
   included in e.g., docs;

 - move std/internal/{transform,generate}-module.js "entry point"
   templates to std/cmd, so they are with their companions;

 - give a @jkcfg/std path when loading the entry points, so they can
   import non-public modules

 - tell typedoc to ignore cmd/, similarly to internal/
